Added [<UseCulture>] attribute,
and used it to explicitly run Swedish tests in the "sv-SV" culture. The
reason for this is that the example "sexhundraarton" contains a double-a,
which on my Danish system (running on da-DK) is interpreted as the letter
`å`.

Annotating all Swedish tests with [<UseCulture("sv-SV")>] solves the
problem, in the sense that all tests now pass on my machine.

This solution ought to be applied to all the other tests as well,
explicitly using the particular culture in use.